% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremeForest.R
\name{erf_cv}
\alias{erf_cv}
\title{Cross validation (CV) for Extreme Random Forest}
\usage{
erf_cv(
  X,
  Y,
  t_xi,
  min.node.size = 5,
  K = 5,
  n_rep = 1,
  args_grf = list(),
  args_erf = list(),
  rng = NULL
)
}
\arguments{
\item{X}{Numeric matrix. Matrix of predictors.}

\item{Y}{Numeric vector. Vector of responses.}

\item{t_xi}{Numeric vector. Vector of thresholds.}

\item{min.node.size}{Numeric vector. Grid of values of \code{min.node.size}
to cross validate.
Default is \code{min.node.size = c(5)}.}

\item{K}{Positive integer. Number of folds in CV.
Default is \code{K = 5}.}

\item{n_rep}{Positive integer. Number of CV repetitions.
Default is \code{n_rep = 1}.}

\item{args_grf}{List. List of named arguments for
\code{\link[grf]{quantile_forest}}.
Default is \code{args_grf = list()}.}

\item{args_erf}{List. List of named arguments for
\code{\link[erf]{predict_erf}}.
Default is \code{args_erf = list()}.}

\item{rng}{Numeric vector. Vector with
\code{K * n_rep * length(min.node.size) + 1} seeds used to
reproduce the random splits and fitting.
Default is \code{rng = NULL}.}
}
\value{
Named list. The list is made of:
\itemize{
\item \code{min.node.size} --- Numeric vector. Grid of values of \code{min.node.size}
to cross validate.
\item \code{cv_error} --- Numeric vector. Vector of cross validation errors.
\item \code{cv_se} --- Numeric vector. Vector of cross validation standard errors.
}
}
\description{
Runs repeated CV for Extreme Random Forest.
}
\references{
\insertAllCited{}
}
