% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_forest.R
\name{extreme_forest}
\alias{extreme_forest}
\title{Fit an extremal random forest (ERF)}
\usage{
extreme_forest(X, Y, min.node.size, lambda, intermediate_estimator, ...)
}
\arguments{
\item{X}{\code{numeric matrix} --- Matrix of predictors, where each row corresponds to
an observation and each column to a predictor.}

\item{Y}{(numeric vector) response vector.}

\item{min.node.size}{\code{numeric >= 0} --- A target for the minimum
number of observations in each tree leaf used in \code{\link[grf:quantile_forest]{grf::quantile_forest()}}.
Nodes with size smaller than \code{min.node.size} can occur,
as in the original \pkg{randomForest} package.
Default is 5.}

\item{lambda}{(numeric \eqn{\geq}{>=} 0)}

\item{intermediate_estimator}{!!! document this}

\item{...}{!!! document this}
}
\value{
An object with S3 class "\code{extreme_forest}",
that is a named list with the following elements.

\item{quantile_forest}{An object with S3 class "\code{quantile_forest}". See
\code{\link[grf:quantile_forest]{grf::quantile_forest()}}.}

\item{intermediate_estimator}{An object with arbitrary S3 class that possesses the
"\code{predict}" method. Such "\code{predict}" method must accept
as \strong{second} argument a numeric matrix with the test predictors, e.g.,
\code{newdata}.
For example, \code{predict(intermediate_estimator, newdata, ...)}
must execute with no errors.}

\item{min.node.size}{!!! document this}

\item{lambda}{!!! document this}
}
\description{
Fit an extremal random forest ... !!! write description
}
\details{
!!! write details
}
\examples{
"!!! add examples"


}
