% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_erf.R
\name{new_erf}
\alias{new_erf}
\title{Fit an extremal random forest (ERF)}
\usage{
new_erf(X, Y, min.node.size, lambda, intermediate_threshold)
}
\arguments{
\item{X}{Numeric matrix (or \code{data.frame}) of predictors, where each row
corresponds to an observation and each column to a predictor.}

\item{Y}{Numeric vector of responses.}

\item{min.node.size}{Minimum number of observations in each tree leaf used
in \code{\link[grf:quantile_forest]{grf::quantile_forest()}}.
Nodes with size smaller than \code{min.node.size} can occur,
as in the original \pkg{randomForest} package.
Default is 5.}

\item{lambda}{Penalty for the shape parameter used in the weighted likelihood.
Default is 0.001.}

\item{intermediate_threshold}{An object with arbitrary S3 class that
possesses a \code{predict} method.
Such "\code{predict}" method must accept a numeric matrix (or \code{data.frame})
of test predictors, e.g., \code{newdata}, as \strong{second} argument.
For example, \code{predict(intermediate_threshold, newdata, ...)}
must execute with no errors.

This is a fitted object used to predict the intermediate thresholds.}
}
\value{
An object with S3 class "\code{erf}".
It is a named list with the following elements:

\item{quantile_forest}{An object with S3 class "\code{quantile_forest}" (see
\code{\link[grf:quantile_forest]{grf::quantile_forest()}}). This is a fitted generalized random forest that
contains the similarity weights used in the log-likelihood.}

\item{min.node.size}{Minimum number of observations in each tree leaf used
to fit the \code{quantile_forest}.}

\item{lambda}{Penalty for the shape parameter used in the weighted
log-likelihood.}

\item{intermediate_threshold}{An object with arbitraty S3 class, as
described in the \strong{Arguments} section.
It is a fitted object used to predict the intermediate thresholds.}
}
\description{
Internal function to fit an extremal random forest ... !!! write description
}
\details{
!!! write details
}
\examples{
"!!! add examples"
}
