% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.erf.R
\name{predict.erf}
\alias{predict.erf}
\title{!!! code \code{predict.erf(object, newdata = NULL, quantiles = NULL, quantile_intermediate = NULL)}
handle errors for
\itemize{
\item \code{object} that is not \code{erf}
\item \code{...}
@export
}}
\usage{
\method{predict}{erf}(
  object,
  newdata = NULL,
  quantiles = c(0.95, 0.99),
  quantile_intermediate = 0.8,
  ...
)
}
\description{
!!! code \code{predict.erf(object, newdata = NULL, quantiles = NULL, quantile_intermediate = NULL)}
handle errors for
\itemize{
\item \code{object} that is not \code{erf}
\item \code{...}
@export
}
}
