% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_forest.R
\name{validate_data}
\alias{validate_data}
\title{!!! code \code{predict.extreme_forest(object, newdata = NULL, quantiles = NULL, quantile_intermediate = NULL)}
handle errors for
\itemize{
\item \code{object} that is not \code{extreme_forest}
\item ...
}}
\usage{
validate_data(X, Y)
}
\description{
!!! code \code{predict.extreme_forest(object, newdata = NULL, quantiles = NULL, quantile_intermediate = NULL)}
handle errors for
\itemize{
\item \code{object} that is not \code{extreme_forest}
\item ...
}
}
