% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erf-package.R
\docType{package}
\name{erf-package}
\alias{erf-package}
\title{erf: Extreme Random Forest}
\description{
Extreme quantile prediction in high-dimensions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nicolagnecco/erf}
  \item Report bugs at \url{https://github.com/nicolagnecco/erf/issues}
}

}
\author{
\strong{Maintainer}: Nicola Gnecco \email{nicola.gnecco@gmail.com}

Authors:
\itemize{
  \item Edossa Terefe
  \item Sebastian Engelke
}

}
\keyword{internal}
