% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erf_cv.R
\name{erf_cv}
\alias{erf_cv}
\title{Cross-validation for an extremal random forest (ERF)}
\usage{
erf_cv(
  X,
  Y,
  min.node.size = c(5, 40, 100),
  lambda = c(0, 0.001, 0.01),
  intermediate_estimator = c("grf", "neural_nets"),
  intermediate_quantile = 0.8,
  nfolds = 5,
  nreps = 3,
  seed = NULL
)
}
\arguments{
\item{X}{Numeric matrix of predictors, where each row
corresponds to an observation and each column to a predictor.}

\item{Y}{Numeric vector of responses.}

\item{min.node.size}{Vector with minimum number of observations in each tree
leaf used to fit the similarity weights
(see also \code{\link[grf:quantile_forest]{grf::quantile_forest()}}).
Nodes with size smaller than \code{min.node.size} can occur,
as in the original \pkg{randomForest} package.
Default is \code{c(5, 40, 100)}.}

\item{lambda}{Vector with penalties for the shape parameter used in the weighted likelihood.
Default is \code{c(0, 0.001, 0.01)}.}

\item{intermediate_estimator}{A character specifying the estimator used to
fit the intermediate threshold.
Options available are:
\itemize{
\item \code{grf}, see \code{\link[grf:quantile_forest]{grf::quantile_forest()}}.
\item \code{neural_nets}, (coming soon).
}}

\item{intermediate_quantile}{Intermediate quantile
level, used to predict the intermediate threshold.
For further information see \insertCite{merg2020;textual}{erf}.
Default is \code{0.8}.}

\item{nfolds}{Number of folds in the cross-validation scheme.
Default is \code{5}.}

\item{nreps}{Number of times \code{nfolds} cross-validation is repeated.
Default is \code{3}.}

\item{seed}{Random seed to reproduce the fold splits.
Default is \code{NULL}.}
}
\value{
An object with S3 class "\code{erf_cv}".
It is a named list with the following elements:

\item{scores}{A \code{tibble} with columns: \code{min.node.size}, \code{lambda}, \code{cvm}
(mean cross-validated error).}

\item{erf}{A fitted "\code{erf}" object on the full data using the optimal
\code{min.node.size} and \code{lambda}.}
}
\description{
Fits an extremal random forest (ERF) with cross-validation.
}
