% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremeForest.R
\name{extreme_forest}
\alias{extreme_forest}
\title{ExtremeForest}
\usage{
extreme_forest(
  X.train,
  Y,
  X.test,
  alpha,
  alpha.new,
  Y.test = NULL,
  grf = FALSE
)
}
\description{
Computes quantile extreme forest ...
}
