% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_forest.R
\name{extreme_forest}
\alias{extreme_forest}
\title{Fit an extremal random forest (ERF)}
\usage{
extreme_forest(
  X,
  Y,
  min.node.size = 5,
  lambda = 0.001,
  intermediate_estimator = c("grf", "neural_nets"),
  ...
)
}
\arguments{
\item{X}{Numeric matrix of predictors, where each row
corresponds to an observation and each column to a predictor.}

\item{Y}{Numeric vector of responses.}

\item{min.node.size}{Minimum number of observations in each tree leaf used
in \code{\link[grf:quantile_forest]{grf::quantile_forest()}}.
Nodes with size smaller than \code{min.node.size} can occur,
as in the original \pkg{randomForest} package.
Default is 5.}

\item{lambda}{Penalty for the shape parameter used in the weighted likelihood.
Default is 0.001.}

\item{intermediate_estimator}{A character specifying the estimator used to
fit the intermediate threshold.
Options available are:
\itemize{
\item \code{grf}, see \code{\link[grf:quantile_forest]{grf::quantile_forest()}}.
\item \code{neural_nets}, (coming soon).
}}

\item{...}{Additional arguments used by the \code{intermediate_estimator}.}
}
\value{
An object with S3 class "\code{extreme_forest}".
It is a named list with the following elements:

\item{quantile_forest}{An object with S3 class "\code{quantile_forest}" (see
\code{\link[grf:quantile_forest]{grf::quantile_forest()}}). This is a fitted generalized random forest that
contains the similarity weights used in the log-likelihood.}

\item{min.node.size}{Minimum number of observations in each tree leaf used
to fit the \code{quantile_forest}.}

\item{lambda}{Penalty for the shape parameter used in the weighted
log-likelihood.}

\item{intermediate_threshold}{An object with S3 class:
\itemize{
\item "\code{quantile_forest}", if \code{intermediate_estimator = "grf"}.
\item "\code{...}", if \verb{intermediate_estimator = "neural_nets}.
}

This is a fitted object used to predict the intermediate thresholds.}
}
\description{
Fit an extremal random forest ... !!! write description
}
\details{
!!! write details
}
\examples{
"!!! add examples"

}
