% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremeForest.R
\name{predict_erf}
\alias{predict_erf}
\title{Extreme predictions with a quantile forest}
\usage{
predict_erf(
  object,
  quantiles,
  threshold = 0.8,
  newdata = NULL,
  model_assessment = FALSE,
  Y.test = NULL,
  out_of_bag = FALSE,
  param_est = c("ML", "Hill"),
  lambda = 0
)
}
\arguments{
\item{object}{Quantile forest object. The trained forest obtained with
\code{\link[grf]{quantile_forest}}.}

\item{quantiles}{Numeric vector (0, 1).
Extreme quantiles at which estimates are required.}

\item{threshold}{Numeric (0, 1). Intermediate quantile used to compute
thresholds \eqn{t(x_i)} for the GPD.
Note that \code{threshold} < \code{quantiles}.
Default is 0.8.}

\item{newdata}{Numeric matrix.
Rows contain observation at which predictions should be made.
If NULL, makes out-of-bag predictions on the training set
instead (i.e., provides predictions at Xi using only trees
that did not use the i-th training example). Note
that this matrix should have the number of columns as the
training matrix, and that the columns must appear in the same
order. This argument is needed if
\code{model_assessment = TRUE}.
Default is \code{NULL}.}

\item{model_assessment}{Boolean. Assess GPD fit with QQ-plot against
exponential distribution?
Default is \code{FALSE}.}

\item{Y.test}{Numeric vector. Responses for the test set, needed
if \code{model_assessment = TRUE}.
Default is \code{NULL}.}

\item{out_of_bag}{Boolean. Use out-of-bag observations to compute thresholds
\eqn{t(x_i)} for the GPD?
Default is \code{FALSE}.}

\item{param_est}{Character. One of "ML" and "Hill". Specifies the method to
fit the shape and scale parameter of the GDP distribution.
"ML" is the weighted maximum likelihood approach proposed by
\insertCite{merg2020;textual}{erf}.
"Hill" is the weighted Hill's estimator approach proposed
by \insertCite{deub2020;textual}{erf}.}

\item{lambda}{Numeric (>= 0). Penalty coefficient for the \eqn{\xi}{csi} parameter
in the weighted log-likelihood, i.e., when \code{param_est = "ML"}.
Default is \code{0}.}
}
\value{
Named list. The list is made of:
\itemize{
\item \code{predictions} --- Numeric matrix. Predictions at each test point (on rows) for
each desired quantile (on columns).
\item \code{pars} --- Numeric matrix. Estimated parameters at each test point (on rows).
The columns contain the estimates for the \eqn{\sigma}{sigma} and
\eqn{\xi}{csi} parameter, respectively.
\item \code{threshold} --- Numeric matrix. Intermediate thresholds at each test point (on rows),
estimated using \code{\link[grf]{quantile_forest}}.
\item \code{plot} (if \code{model_assessment = TRUE}). QQ-plot for model assessment.
}
}
\description{
Predicts high conditional quantiles of Y given X using a quantile forest
fitted using \code{\link[grf]{quantile_forest}}.
}
