% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremeForest.R
\name{predict_erf}
\alias{predict_erf}
\title{Extreme predictions with a quantile forest}
\usage{
predict_erf(
  object,
  quantiles,
  threshold = 0.8,
  newdata = NULL,
  model_assessment = FALSE,
  Y.test = NULL,
  out_of_bag = FALSE
)
}
\arguments{
\item{object}{Quantile forest object. The trained forest.}

\item{quantiles}{Numeric vector (0, 1).
Extreme quantiles at which estimates are required.}

\item{threshold}{Numeric (0, 1). Intermediate quantile used to compute
thresholds \eqn{t(x_i)} for the GPD.
Note that \code{threshold} < \code{quantiles}.
Default is 0.8.}

\item{newdata}{Numeric matrix.
Rows contain observation at which predictions should be made.
If NULL, makes out-of-bag predictions on the training set
instead (i.e., provides predictions at Xi using only trees
that did not use the i-th training example). Note
that this matrix should have the number of columns as the
training matrix, and that the columns must appear in the same
order. This argument is needed if
\code{model_assessment = TRUE}.
Default is \code{NULL}.}

\item{model_assessment}{Boolean. Assess GPD fit with QQ-plot against
exponential distribution?
Default is \code{FALSE}.}

\item{Y.test}{Numeric vector. Responses for the test set, needed
if \code{model_assessment = TRUE}.
Default is \code{NULL}.}

\item{out_of_bag}{Boolean. Use out-of-bag observations to compute thresholds
\eqn{t(x_i)} for the GPD?
Default is \code{FALSE}.}
}
\value{
List. The list is made of:
\enumerate{
\item Numeric matrix. Predictions at each test point (on rows) for
each desired quantile (on columns).
\item Plot if \code{model_assessment = TRUE}. QQ-plot for model assessment.!!! return also t(x_0), sigma(x_0), csi(x_0), for all x_0 in test_data

}
}
\description{
Gets estimates of the extrapolated conditional quantiles of Y given X
using a trained forest.
}
